download

+++ package/chemicalchecker/core/proj.py
@@ -73,7 +73,7 @@ class proj(BaseSignature, DataSignature):
                 dst.create_dataset("mappings", data=src['mappings'][:],
                                    dtype=sdtype)
             src_len = src["V"].shape[0]
-            dst.create_dataset("V", (src_len, n_components), dtype=np.float32)
+            dst.create_dataset("V", (src_len, n_components), dtype='float32')
             for i in tqdm(range(0, src_len, batch_size), 'PRE transform'):
                 chunk = slice(i, i + batch_size)
                 dst['V'][chunk] = preprocess_algo.transform(src['V'][chunk])
@@ -100,7 +100,7 @@ class proj(BaseSignature, DataSignature):
                 dst.create_dataset("mappings", data=src['mappings'][:],
                                    dtype=sdtype)
             src_len = src["V"].shape[0]
-            dst.create_dataset("V", (src_len, 2), dtype=float)
+            dst.create_dataset("V", (src_len, 2), dtype='float32')
             for i in tqdm(range(0, src_len, batch_size), 'transform'):
                 chunk = slice(i, i + batch_size)
                 dst['V'][chunk] = preprocess_algo.transform(src['V'][chunk])
